Main

Description

Módulo Haskell para a Tarefa D.

Documentation

type Nivel = [String]

O tipo Nivel, representa uma String de Strings, ou seja, as linhas do mapa, e as coordenadas e comandos.

type Mapa = [Linha]

O tipo Mapa, representa uma String das Linhas do mapa do jogo.

type Coordenadas = String

O tipo Coordenadas, é dado por uma String.

type Comandos = String

O tipo Comandos, é dado por uma String.

type Resultado = String

O tipo Resultado, é dado por uma String, representa o código completo, para apresentar a animação 3D do nível.

type Linha = String

O tipo Linha, respresenta uma linha do mapa, numa String.

type Orientacao = Char

O tipo Orientacao, representa a orientação do robo no mapa, que é dado num Char.

type Posicao = (Int, Int, Char)

o tipo Posicao, respresenta o valor de x,y e a orientação num tuplo.

type CoordLampada = [Int]

O tipo CoordLampada, representa a lista de Int das coordenadas de uma lâmpada.

main :: IO ()

testa :: FilePath -> IO ()

tarefa :: Nivel -> Resultado

A função tarefa, recebe o Nivel do jogo, e fornece a função tarefa2, os dados referente ao Nivel, como a posição em tuplo e a lista de lista de Int, com as coordenadas das lâmpadas.

Exemplo de utilização :

>>> tarefa ["aaa","aAa","aaa","0 0 N"]
"ADAL"

tarefa2 :: Nivel -> Posicao -> [CoordLampada] -> Comandos

A função tarefa2, recebe um Nivel, uma posição e uma lista de lista de Int, das coordendas das lâmpadas. Esta função respresenta o motor da tarefa E, devido ao facto de ser a junção de todas as funções abaixo elaboradas. Esta função, aplica a função igualY , a função igualX, de forma a que o robô se deloca em L, para as lâmpadas. No entanto, caso nenhumas destas formas sejam possiveis, a tarefa muda a ordem da lista das lâmpadas a acenter. Por fim, como último caso, temos a função parede, para caso do aparecimento de obstáculos.

Exemplo de utilização :

>>> tarefa2 ["Aaa","aAa","aaa","0 0 N"] (0,0,'N') [[2,0]]
"ADALEAEAL"

parede :: Nivel -> Posicao -> [CoordLampada] -> Comandos

A função parede, recebe um Nivel, uma posição do robô e a lista de lista de Int, das coordenadas das lâmpadas. Esta função, permite ultrapassar obstáculos, ou seja, verificar se a próxima posição é válida para avançar ou nâo, devido ao obstáculos. Caso haja um obstáculo, a Norte do robô, a função desloca-o uma casa para Este e verifica se o robô pode avançar, através da função tarefa2. Caso o obstáculo se encontra a Este, ele desloca-se uma casa para Norte, e verifica se pode avançar. Para as outras orientações, o robô, ajusta-se em função do Norte e do este. Concluindo esta função, perimite ao robô, desviar-se do osbtáculos presentes no Nível.

Exemplo de utilização :

>>> parede ["Aaa","caa","aaa","0 0 N"] (0,0,'N') [[2,0]]
"DAEAAEAL"

avan :: Mapa -> Posicao -> Posicao -> [CoordLampada] -> Comandos

A função avan, recebe um Mapa, duas posiçôes, sendo a primeira a inical e a segunda a seguinte, e a lista de lista de Int, das coordenadas das lâmpadas. A função tem como objectivo, validar, se o robô pode ou não avançar, e se caso pode avançar, verifica se anda ou salta. Caso nao possa avançar, devolve "1", para que seja detéctavél noutra função.

Exemplo de utilização :

>>> avan ["aaA","aAa","aaa"] (0,1,'N') (0,1,'N') [[1,1],[2,2]]
"A"

igualY :: Mapa -> Posicao -> CoordLampada -> Comandos

A função igualY, recebe um Mapa, uma Posicao do robô, e uma lista de Int, das coordenadas de uma lâmpada. Esta função compara o valor da cooordenada y da posição do robô ,com a coordenada y da lâmpada, e consoante for maior ou menor, vai aplicar alerações ao estado do robô, rodando para a esquerda ou para a direita, conforme a situação, ou caso seja possível, avança ou salta. A função tem em atenção orientar o robô, para Norte ou para Sul, para permitir que o robô possa avançar, tendo em conta a coordenada y, isto é, se a coordenada y for maior do robô do que a da lâmpada, a função roda o robô para Sul, e caso seja o contrário, roda-o para Norte. É de destacar, que a função quando obtem o mesmo valor em y, devolve a orientação do robô, para que possa ser ultilizada por outra função. Caso nenhum dos parámetros sejam aceitaveis, a função devolve "1", uma forma de detéctar caso seja a função aplicavél ou não.

Exemplo de utilização :

>>> igualY ["aaa","aAa","aaa"] (0,1,'N') [1,1]
"EAN"

igualX :: Mapa -> Posicao -> CoordLampada -> Comandos

A função igualX, em semelhança a função igualY, recebe um Mapa, uma Posicao do robô, e uma lista de Int, das coordenadas de uma lâmpada. Esta função atua de forma identica a função igualY, no entanto, difere a seu objectivo. Ao contrário da função igualY, esta funciona para a coordenada y, tentando orientar o robô, para Oeste caso a coordenada x do robô seja maior do que da lâmpada, ou para este, caso contrário.

Exemplo de utilização :

>>> igualX ["aaa","aAa","aaa"] (0,0,'N') [1,1]
"DAE"

trans :: Char -> Char

A função trans, recebe uma posição e tranforma, caso for um caratér entre A e Z, em caractéres entre a e z. Esta função permite que os caractéres de A até Z, sejam comparáveis com os caractéres de a até z.

Exemplo de utilização :

>>> trans 'A'
'a'

troca :: [CoordLampada] -> [CoordLampada]

A função troca, recebe uma lista de listas de Int, que representam as coordenadas das lâmpadas. Desta forma, a função, passa a primeira lista, para o fim da lista, o que permite na função principal, o robô, deslocar-se para outra lâmpada, caso a primeira da lista seja dificil de atingir. sejam comparáveis com os caractéres de a até z.

Exemplo de utilização :

>>> troca [[1,1],[2,2]]
[[2,2],[1,1]]

localL1 :: Linha -> Int -> Int -> [CoordLampada]

A função localL1 recebe uma Linha do Mapa, e devolve as coordenadas,numa lista de lista de Int, dos pontos que possuiem lâmpadas, isto é, as posições que tem letras maiúsculas.

Exemplo de utilização :

>>> localL1 "acbBba" 0 0
[[3,0]] 

localL :: Mapa -> Int -> Int -> [CoordLampada]

A função localL recebe uma Mapa, e devolve todas as coordenadas,numa lista de lista de Int, dos pontos que possuiem lâmpadas, isto é, as posições que tem letras maiúsculas.

Exemplo de utilização :

>>> localL ["acbBba","bbbAab"] 0 0
[[3,0],[3,1]] 

pos :: Coordenadas -> Posicao

A função robô, recebe uma String, nomeadamente, as Coordendas presente no Nivel, e passa para um tuplo, de forma a obter a coordenada x, em primeiro, a coordenada y, em segundo, e a orientação em último.

Exemplo de utilização :

>>> pos "2 1 N"
(2,1,'N')

posicao :: Mapa -> Int -> Int -> Char

A função posicao, recebe o Mapa, e dois Int, que representam as coordenadas x e y. Esta função dá-nos o local exacto do robô no mapa, devolvendo assim o Char em que se localiza.

Exemplo de utilização :

>>> posicao ["aBac","aAca","aabb","2 1 N","ASDAA"]
"c"

posicaoSe :: Mapa -> Int -> Int -> Orientacao -> Char

A função posicaoSe, recebe o Mapa, dois Int, que representam as coordenadas x e y, e um Char, ou seja, a orientação do robô. Esta função dá-nos o local da posição seguinte, consoante a orientação do robô. Devolve assim o Char da próxima posição.

Exemplo de utilização :

>>> posicao ["aBac","aAca","aabb","2 1 N","ASDAA"]
"a"

strInt :: String -> Int

A função strInt recebe uma String e converte em um Int correspodente, com recurso a função auxiliar converte.

Exemplo de utilização :

>>> strInt "123"
123

converte :: Int -> [Int] -> Int

A função converte recebe um Int e uma lista de Int. Esta função passa a lista de Int, para o respectivo número, calculando-o, tendo em conta a base 10. O primeiro argumento, Int, serve como um acumulador para a base 10.

Exemplo de utilização :

>>> converte 0 [1,2,3]
123