Main

Description

Módulo Haskell para a Tarefa E.

Documentation

main :: IO ()

testa :: FilePath -> IO ()

func :: IO ()

A função func é uma função auxiliar, que na realização deste módulo Haskell, permitia passar o output da função principal para um ficheiro, que se tornava numa forma fácil de poder verificar se os outputs estavam a realizar-se de forma correta.

type Nivel = [String]

O tipo Nivel, representa uma String de Strings, ou seja, as linhas do mapa, e as coordenadas e comandos.

type Mapa = [Linha]

O tipo Mapa, representa uma String das Linhas do mapa do jogo.

type Coordenadas = String

O tipo Coordenadas, é dado por uma String.

type Comandos = String

O tipo Comandos, é dado por uma String.

type Resultado = String

O tipo Resultado, é dado por uma String, representa o código completo, para apresentar a animação 3D do nível.

type Linha = String

O tipo Linha, respresenta uma linha do mapa, numa String.

type Orientacao = Char

O tipo Orientacao, representa a orientação do robo no mapa, que é dado num Char.

type Vx = Int

o tipo Vx, respresenta o valor de x.

type Vy = Int

o tipo Vy, respresenta o valor de y.

type Vz = Int

o tipo Vz, respresenta o valor de z, para a representação em 3D.

type Posicao = Char

O tipo Posicao, representa a letra onde se situa o robo no mapa.

type Html = [String]

O tipo Html, representa o código html apresentado em uma lista de String.

type LinhaHtml = String

O tipo LinhaHtml, representa uma linha de código html apresentado em uma String.

type CoordLampadas = String

O tipo CoordLampadas, representa a String, com as coordenadas das lâmpadas.

tarefa :: Nivel -> Resultado

A função tarefa é a função principal, pois reúne todas as funções elaboradas ao longo deste módulo Haskell, tendo como objectivo reproduzir um código html, onde este apresenta uma animação 3D, com base em x3dom.

formaTab :: Linha -> Vx -> Vy -> Html

A função formaTab, é uma função que recebe uma linha do mapa, dois Int, servindo estes como acumuladores, dando as coordenadas no decorrer da função. A função devolve um código html, que apresenta uma linha do tabuleiro do jogo, sem as lâmpadas. A apresentação tem em conta, para casos em que o mapa tenha diferentes alturas, apresenta os cubos respectivos por baixo, de forma a que o tabuleiro esteja apresentavél.

Exemplo de utilização :

>>> formaTab "aa" 0 0
["<Transform translation='0 0 0'> <Shape USE=\"tile\"/> </Transform>","<Transform translation='1 0 0'> <Shape USE=\"tile\"/> </Transform>"]

formaTab2 :: Mapa -> Vx -> Vy -> Html

A função formaTab2, é uma função que aplica de forma recursiva a função formaTab, precorrendo todas as linhas do mapa, devolvendo um código html, com o tabuleiro completo, no entanto, sem as lâmpadas.

Exemplo de utilização :

>>> formaTab2 ["aa","aa"] 0 0
["<Transform translation='0 0 0'> <Shape USE=\"tile\"/> </Transform>","<Transform translation='1 0 0'> <Shape USE=\"tile\"/> </Transform>",
"<Transform translation='0 1 0'> <Shape USE=\"tile\"/> </Transform>","<Transform translation='1 1 0'> <Shape USE=\"tile\"/> </Transform>"]

formaRobo :: Nivel -> LinhaHtml

A função formaRobo, recebe um Nivel, do qual retira se a posição do robo, e a sua respetiva altura. Desta forma, a função devolve como output, uma linha de código html que apresenta o robô na posição inicial.

Exemplo de utilização :

>>> formaRobo ["aAa","aAa","aAa","0 0 N","AADALDALAL"]
"<Transform scale=\".3 .3 .3\" def=\"robot\" translation='0 0 1' rotation='1 0 0 0'>"

formaOrient :: Coordenadas -> LinhaHtml

A função formaOrient, recebe as Coordenadas da posição inical do robô, e retira a orientação do robô. Desta forma para cada orientação, faz corresponder uma porção de código html que faz orientar o robô em 3D.

Exemplo de utilização :

>>> formaRobo "0 0 N"
"1 0 0 0"

formaRotacao :: Nivel -> Comandos -> Orientacao -> LinhaHtml

A função formaRotacao, recebe as Coordenadas da posição inical do robô, e retira a orientação do robô. Desta forma para cada orientação, faz corresponder uma porção de código html que faz orientar o robô em 3D.

Exemplo de utilização :

>>> formaRotacao "0 0 N"
"1 0 0 0"

formaCam :: Nivel -> Comandos -> Vx -> Vy -> Vz -> Orientacao -> CoordLampadas -> LinhaHtml

A função formaCam, recebe o Nivel, os comandos, coordenadas x, y e z da posição inical do robô, a orietação do robô e a String das coordenadas das lâmpadas. A partir destes argumentos, a função vai ler recursivamente os comandos, aplicando as devias alterações nas coordenadas, tendo em atenção se caso salta para uma posição mais alta ou mais baixa. Caso o comando for L, as coordenadas em que o robô se situa, passam para a String das coordenadas das lampadas.

Exemplo de utilização :

>>> formaCam ["aaa","aaa","aAa","0 0 N","ADAL"] (last ["aaa","aaa","aAa","0 0 N","ADAL"]) 0 0 1 'N' []
"0 1 1 0 1 1 1 1 1 1 1 1 R 1 1 1 "

trans :: Posicao -> Posicao

A função trans, recebe uma posição e tranforma, caso for um caratér entre A e Z, em caractéres entre a e z. Esta função permite que os caractéres de A até Z, sejam comparáveis com os caractéres de a até z.

Exemplo de utilização :

>>> trans 'A'
'a'

retira :: LinhaHtml -> LinhaHtml

A função retira, recebe uma String de uma linha de código html, da função formaCam, e retira a parte referente as coordenadas das lâmpadas.

Exemplo de utilização :

>>> retira "0 1 1 0 1 1 1 1 1 1 1 1 R 1 1 1 "
"0 1 1 0 1 1 1 1 1 1 1 1 "

coordLuz :: [CoordLampadas] -> [Int]

A função coordLuz, recebe uma lista de uma String das coordenadas das lâmpadas, repartida em cada caractér, devido a função words, e retira os caratéres R, e tranforma as coordenadas, numa lista de Int.

Exemplo de utilização :

>>> coordLuz ["R","1","1","1"]
[1,1,1]

formaLamp :: [Int] -> Int -> Html

A função formaLamp, recebe uma lista de Int, proveniente da função coordLuz, e um Int, servindo de acumulador. A função devolve um código html, que permite a formação dos cubos que representam as lâmpadas.

Exemplo de utilização :

>>> formaLamp [1,1,1] 1
["<Shape DEF=\"tileL1\">","<Appearance>","<Material def=\"corLam1\" diffuseColor='0 0 1.0'/>","</Appearance>",
"<Box size='.98 .98 .98'/>","</Shape>","<Transform translation='1 1 0'> <Shape USE=\"tileL1\"/> </Transform>"]

formaLuz :: Comandos -> Comandos -> Int -> Int -> Html

A função formaLuz, recebe duas listas dos Comandos, dois Int, servindo como acumuladores. Esta função devolve um código html, que permite, com auxilio da função temp2, para calular o instante em que a lâmapada liga, a animação 3D em que o robô liga as lâmpadas. O acumulador x, serve para cada lâmpada que surge, criar um DEF diferente, tornando as lâmpadas independentes. O acumulador y, serve para indicar em que altura é aplicado o comando L.

Exemplo de utilização :

>>> formaLuz "ADAL" "ADAL" 1 0
["<PositionInterpolator DEF=\"corL1\" key=\"0 0.75 1.0 1\" keyValue=\"0.57 0.57 0.57  0.57 0.57 0.57  1 1 0  1 1 0\"> </PositionInterpolator>   ",
"<Route fromNode=\"time\" fromField =\"fraction_changed\" toNode=\"corL1\" toField=\"set_fraction\"> </Route> ",
"\t<Route fromNode=\"corL1\" fromField =\"value_changed\" toNode=\"corLam1\" toField=\"diffuseColor\"> </Route>  ",""]

temp :: Comandos -> Comandos -> Int -> LinhaHtml

A função temp recebe duas String e um Int, e devolve uma porção de código html. Esta porção é referente aos intervalos de tempo de cada comando, ou seja, admitindo 1 como o tempo de execução do nível, divindindo pelo numero de comandos, obtemos os intervalos.

Exemplo de utilização :

>>> temp "ADAL" "ADAL" 1
"0.25 0.5 0.75 1.0 "

temp2 :: Comandos -> Comandos -> Int -> LinhaHtml

A função temp2 recebe duas String e um Int, e devolve uma porção de código html. Esta porção é referente um preciso intrevalo de tempo, ou seja, tendo vários comandos, o valor de x vai ser o número de coordenadas até lá aplicadas. Assim a função calcula esse instante e o o próximo apenas.

Exemplo de utilização :

>>> temp2 "ADAL" "ADAL" 2
"0.5 0.75"

divid :: Int -> Int -> Float

A função divid recebe dois Int, e devolve um Float. Esta função permite apresentar em casa decimais, as divisões necessarias na função temp e temp2.

Exemplo de utilização :

>>> divid 1 2
0.5

formaRota :: Nivel -> Html

A função formaRota, recebe o Nivel, e aplica a função temp, que define os intervalos de tempo, a função formaCam, que apresenta a sequência de movimentos do robô, a função formaRotacao, que vai dando a sequências de rotações do robô, e a função formaLuz, que vai dar as instâncias em que cada luz se liga. Devolve assim, um código html, que permite a animação do robô e das lâmpadas ao longo do nível.

Exemplo de utilização :

>>> formaRota ["aaa","aAa","aaa","0 0 N","ADAL"]
["\t<timeSensor DEF=\"time\" cycleInterval=\"4\" loop=\"true\"> </timeSensor>","",
"<PositionInterpolator DEF=\"move\" key=\"0 0.25 0.5 0.75 1.0\" keyValue=\"0 0 1 0 1 1 0 1 1 1 1 1 1 1 1 \"> </PositionInterpolator>       ",
"<Route fromNode=\"time\" fromField =\"fraction_changed\" toNode=\"move\" toField=\"set_fraction\"> </Route> ",
"\t<Route fromNode=\"move\" fromField =\"value_changed\" toNode=\"robot\" toField=\"translation\"> </Route>     ",
"",
"<orientationInterpolator DEF=\"moveR\" key=\"0 0.25 0.5 0.75 1.0\" keyValue=\"1 0 0 0  1 0 0 0  1 1 0 3.14  1 1 0 3.14  1 1 0 3.14\"> </orientationInterpolator> ",
"<Route fromNode=\"time\" fromField=\"fraction_changed\" toNode=\"moveR\" toField=\"set_fraction\"> </Route>",
"<Route fromNode=\"moveR\" fromField=\"value_changed\" toNode=\"robot\" toField=\"rotation\"> </Route>   ",
"",
"<PositionInterpolator DEF=\"corL1\" key=\"0 0.75 1.0 1\" keyValue=\"0.57 0.57 0.57  0.57 0.57 0.57  1 1 0  1 1 0\"> </PositionInterpolator>   ",
"<Route fromNode=\"time\" fromField =\"fraction_changed\" toNode=\"corL1\" toField=\"set_fraction\"> </Route> ",
"\t<Route fromNode=\"corL1\" fromField =\"value_changed\" toNode=\"corLam1\" toField=\"diffuseColor\"> </Route>  ",""]

conve :: Posicao -> Int

A função conve recebe um Char e converte num Int. Esta função permite determinar a altura das posições, devolvendo os valores da coordenada z.

Exemplo de utilização :

>>> conve 'A'
97
>>> conve 'b'
98

posicao :: Mapa -> Vx -> Vy -> Posicao

A função posicao, recebe o Mapa, e dois Int, que representam as coordenadas x e y. Esta função dá-nos o local exacto do robô no mapa, devolvendo assim o Char em que se localiza.

Exemplo de utilização :

>>> posicao ["aBac","aAca","aabb","2 1 N","ASDAA"]
"c"

posicaoSe :: Mapa -> Vx -> Vy -> Orientacao -> Posicao

A função posicaoSe, recebe o Mapa, dois Int, que representam as coordenadas x e y, e um Char, ou seja, a orientação do robô. Esta função dá-nos o local da posição seguinte, consoante a orientação do robô. Devolve assim o Char da próxima posição.

Exemplo de utilização :

>>> posicao ["aBac","aAca","aabb","2 1 N","ASDAA"]
"a"

yponto :: [String] -> Int

A função yponto recebe a String do Nível, e calcula a coordenada Y.

Exemplo de utilização :

>>> xponto  ["aaca","abba","adaa","0 1 N","ASDAA"]
1

xponto :: [String] -> Int

A função xponto recebe a String do Nível, e calcula a coordenada X.

Exemplo de utilização :

>>> xponto  ["aaca","abba","adaa","0 1 N","ASDAA"]
0

ajuda3 :: String -> String

A função ajuda3 recebe uma String, e devolve os primeiros algarismos que encontrar. Esta função auxiliar tem como objectivo, retirar das Coordenadas, o valor de X.

Exemplo de utilização :

>>> ajuda3  "0 1 N"
"0"

ajuda5 :: String -> String

A função ajuda5 recebe uma String, e retira os primeiros algarismos que encontrar, e devolve os próximos algarismos. Esta função auxiliar tem como objectivo, retirar das Coordenadas, o valor de Y.

Exemplo de utilização :

>>> ajuda5  "0 1 N"
"1"

strInt :: String -> Int

A função strInt recebe uma String e converte em um Int correspodente, com recurso a função auxiliar converte.

Exemplo de utilização :

>>> strInt "123"
123

converte :: Int -> [Int] -> Int

A função converte recebe um Int e uma lista de Int. Esta função passa a lista de Int, para o respectivo número, calculando-o, tendo em conta a base 10. O primeiro argumento, Int, serve como um acumulador para a base 10.

Exemplo de utilização :

>>> converte 0 [1,2,3]
123